+++ spring-test/src/test/java/org/springframework/test/context/junit4/InternalSpringJUnit4ClassRunnerTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2007 the original author or authors.
+ * Copyright 2002-2013 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,11 +25,11 @@ import org.springframework.test.context.TestContextManager;
  * @author Sam Brannen
  * @since 2.5
  */
-public class SpringJUnit4ClassRunnerTests {
+public class InternalSpringJUnit4ClassRunnerTests {
 
 	@Test(expected = Exception.class)
 	public void checkThatExceptionsAreNotSilentlySwallowed() throws Exception {
-		SpringJUnit4ClassRunner runner = new SpringJUnit4ClassRunner(getClass()) {
+		InternalSpringJUnit4ClassRunner runner = new InternalSpringJUnit4ClassRunner(getClass()) {
 
 			@Override
 			protected TestContextManager createTestContextManager(Class<?> clazz) {
